moved query map location
