+++ splatnet3_scraper/query/query.py
@@ -1,5 +1,4 @@
 import json
-from typing import ParamSpec, TypeVar
 
 import requests
 
@@ -8,12 +7,8 @@ from splatnet3_scraper.auth.exceptions import SplatNetException
 from splatnet3_scraper.auth.graph_ql_queries import queries
 from splatnet3_scraper.constants import TOKENS
 from splatnet3_scraper.query.config import Config
-from splatnet3_scraper.query.query_map import QueryMap
 from splatnet3_scraper.query.responses import QueryResponse
 from splatnet3_scraper.utils import retry
-
-T = TypeVar("T")
-P = ParamSpec("P")
 
 
 class SplatNet_Query:
