+++ splatnet3_scraper/scraper/query_map.py
